drfecommerce/requirements.txt
asgiref==3.5.2
attrs==23.2.0
black==24.4.2
click==8.1.7
coverage==7.5.1
Django==4.1.4
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
factory-boy==3.3.0
Faker==25.0.1
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.11.0
uritemplate==4.1.1
